feat: refactor the full code
