+++ app/helpers/communication.ts
