+++ app/helpers/index.ts
@@ -1,2 +1,3 @@
 export * from './lookupErrorText';
 export * from './journeyFinder';
+export * from './communication';
