+++ app/models/processingStatements.ts
@@ -1,4 +1,4 @@
-import { get } from "~/utils";
+import { get } from "~/helpers";
 import { IDashboardPSData } from "~/types";
 import CONFIG from "~/config";
